validation: creating a validtation listener manager
