+++ source/extensions/listener_managers/validation_listener_manager/validation_listener_manager.cc
@@ -1,0 +1,9 @@
+#include "source/extensions/listener_managers/validation_listener_manager/validation_listener_manager.h"
+
+namespace Envoy {
+namespace Server {
+
+REGISTER_FACTORY(ValidationListenerManagerFactoryImpl, ListenerManagerFactory);
+
+} // namespace Server
+} // namespace Envoy
