+++ source/server/config_validation/server.h
@@ -24,9 +24,6 @@
 #include "source/common/runtime/runtime_impl.h"
 #include "source/common/secret/secret_manager_impl.h"
 #include "source/common/thread_local/thread_local_impl.h"
-
-// TODO(alyssawilk) address
-#include "source/extensions/listener_managers/listener_manager/listener_manager_impl.h"
 #include "source/server/config_validation/admin.h"
 #include "source/server/config_validation/api.h"
 #include "source/server/config_validation/cluster_manager.h"
@@ -136,58 +133,6 @@ public:
     http_context_.setDefaultTracingConfig(tracing_config);
   }
   void setSinkPredicates(std::unique_ptr<Stats::SinkPredicates>&&) override {}
-
-  class ValidationListenerComponentFactory : public ListenerComponentFactory {
-  public:
-    ValidationListenerComponentFactory(ValidationInstance& parent) : parent_(parent) {}
-
-    // Server::ListenerComponentFactory
-    LdsApiPtr createLdsApi(const envoy::config::core::v3::ConfigSource& lds_config,
-                           const xds::core::v3::ResourceLocator* lds_resources_locator) override {
-      return std::make_unique<LdsApiImpl>(
-          lds_config, lds_resources_locator, parent_.clusterManager(), parent_.initManager(),
-          parent_.stats(), parent_.listenerManager(),
-          parent_.messageValidationContext().dynamicValidationVisitor());
-    }
-    std::vector<Network::FilterFactoryCb> createNetworkFilterFactoryList(
-        const Protobuf::RepeatedPtrField<envoy::config::listener::v3::Filter>& filters,
-        Server::Configuration::FilterChainFactoryContext& filter_chain_factory_context) override {
-      return ProdListenerComponentFactory::createNetworkFilterFactoryListImpl(
-          filters, filter_chain_factory_context);
-    }
-    Filter::ListenerFilterFactoriesList createListenerFilterFactoryList(
-        const Protobuf::RepeatedPtrField<envoy::config::listener::v3::ListenerFilter>& filters,
-        Configuration::ListenerFactoryContext& context) override {
-      return ProdListenerComponentFactory::createListenerFilterFactoryListImpl(
-          filters, context, parent_.tcp_listener_config_provider_manager_);
-    }
-    std::vector<Network::UdpListenerFilterFactoryCb> createUdpListenerFilterFactoryList(
-        const Protobuf::RepeatedPtrField<envoy::config::listener::v3::ListenerFilter>& filters,
-        Configuration::ListenerFactoryContext& context) override {
-      return ProdListenerComponentFactory::createUdpListenerFilterFactoryListImpl(filters, context);
-    }
-    Network::SocketSharedPtr
-    createListenSocket(Network::Address::InstanceConstSharedPtr, Network::Socket::Type,
-                       const Network::Socket::OptionsSharedPtr&, ListenerComponentFactory::BindType,
-                       const Network::SocketCreationOptions&, uint32_t) override {
-      // Returned sockets are not currently used so we can return nothing here safely vs. a
-      // validation mock.
-      // TODO(mattklein123): The fact that this returns nullptr makes the production code more
-      // convoluted than it needs to be. Fix this to return a mock in a follow up.
-      return nullptr;
-    }
-    DrainManagerPtr createDrainManager(envoy::config::listener::v3::Listener::DrainType) override {
-      return nullptr;
-    }
-    uint64_t nextListenerTag() override { return 0; }
-    Filter::TcpListenerFilterConfigProviderManagerImpl*
-    getTcpListenerConfigProviderManager() override {
-      return &parent_.tcp_listener_config_provider_manager_;
-    }
-
-  private:
-    ValidationInstance& parent_;
-  };
 
   // Server::WorkerFactory
   WorkerPtr createWorker(uint32_t, OverloadManager&, const std::string&) override {
